agentAI/requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
plotly
openpyxl
xlrd
statsmodels
requests
polars
cloudpickle
lightgbm
python-dotenv
flask_swagger_ui
flask_swagger
jsonify
flask
python-dotenv
pymupdf
PyPDF2